VFS v3: stat, mkdir, directory listings
