+++ libretro-common/lists/dir_list.c
@@ -171,12 +171,10 @@ static int dir_list_read(const char *dir,
       bool include_dirs, bool include_hidden,
       bool include_compressed, bool recursive)
 {
-   struct RDIR *entry = retro_opendir(dir);
+   struct RDIR *entry = retro_opendir_include_hidden(dir, include_hidden);
 
    if (!entry || retro_dirent_error(entry))
       goto error;
-
-   retro_dirent_include_hidden(entry, include_hidden);
 
    while (retro_readdir(entry))
    {
